flash/FLM.c: Add support for Arm FLM download algorithms
Many manufacturers provides FLM download algorithms. This patch makes it possible to use openocd on new devices without modifying openocd.
An example target file (flm_stm32f4_512k) is also included, the FLM file is not included, so it needs to be copied from device pack.
This patch uses a file from Arm (FlashOS.h), it's licensed under Apache license. I'm not sure if it's ok to use it.

Signed-off-by: Tianyi Wang <163.wty@163.com>
Change-Id: Ie7f5d99ff9375f7c86233163f4f79bc9d7f3f532